hft-marl-phase4_7/requirements.txt
numpy>=1.26
pandas>=2.2
pyarrow>=15.0
scipy>=1.11
pyyaml>=6.0
matplotlib>=3.8
torch>=2.2
gymnasium>=0.29
mlflow>=2.12
abides-markets>=1.8
jax>=0.4.28
jaxlib>=0.4.28
# NOTE: JAX-LOB may require manual install from source; keep optional.
